feat: 검색결과 타입을 런타임에 체크하기 위해 태그된 유니온 기법 사용(#112)
